boinc: separate package for wrapper
The BOINC wrapper is typically distributed by the
BOINC project. But just like any of the scientific
applications, the compute-time contributors can
substitute such an app with their own version.

This package provides a dynamically linked wrapper
that may substitute the effort of the project maintainer
to provide a static version of the same.

Signed-off-by: Steffen Möller <moeller@debian.org>